--- a/movingmedian_test.go
+++ b/movingmedian_test.go
@@ -1,10 +1,10 @@
 package movingmedian
 
 import (
-	"github.com/wangjohn/quickselect"
 	"log"
 	"math"
 	"math/rand"
+	"sort"
 	"testing"
 )
 
@@ -186,34 +186,23 @@ func median(data []float64, i, windowSize int) float64 {
 		min = 0
 	}
 
-	window := make([]float64, 1+i-min)
-	copy(window, data[min:i+1])
-	return percentile(window, 50, true)
-}
-
-func percentile(data []float64, percent float64, interpolate bool) float64 {
-	if len(data) == 0 || percent < 0 || percent > 100 {
+	if len(data) == 0 {
 		return math.NaN()
 	}
-	if len(data) == 1 {
-		return data[0]
+
+	window := make([]float64, 1+i-min)
+	copy(window, data[min:i+1])
+
+	if len(window) == 1 {
+		return window[0]
 	}
 
-	k := (float64(len(data)-1) * percent) / 100
-	length := int(math.Ceil(k)) + 1
-	quickselect.Float64QuickSelect(data, length)
-	top, secondTop := math.Inf(-1), math.Inf(-1)
-	for _, val := range data[0:length] {
-		if val > top {
-			secondTop = top
-			top = val
-		} else if val > secondTop {
-			secondTop = val
-		}
+	sort.Float64s(window)
+
+	k := len(window) / 2
+	if len(window)%2 == 1 {
+		return window[k]
 	}
-	remainder := k - float64(int(k))
-	if remainder == 0 || !interpolate {
-		return top
-	}
-	return (top * remainder) + (secondTop * (1 - remainder))
+
+	return 0.5*window[k-1] + 0.5*window[k]
 }