Random tests are now passing.

There were two bugs that needed fixing.  One was resetting the itemPtr.idx to 0, the other was the missing test in the 'if'.