@dgryski 2015-06-15
This PR oddly only affects performance in the (rare) windowsize=1 case.

```
<dgryski@kaepora[movingmedian] \ʕ◔ϖ◔ʔ/ > benchstat old.txt new.txt 
name                       old time/op  new time/op  delta
_10values_windowsize1       792ns ±12%   841ns ±12%  +6.12%  (p=0.002 n=19+20)
_100values_windowsize10    15.2µs ±18%  15.7µs ± 9%    ~     (p=0.052 n=20+20)
_10Kvalues_windowsize100   2.75ms ± 9%  2.80ms ± 7%    ~     (p=0.126 n=18+20)
_10Kvalues_windowsize1000  4.00ms ±10%  3.94ms ± 7%    ~     (p=0.531 n=20+19)
```