@dgryski 2015-06-15
I'll reduce the maximums.  I just wanted to be sure to hit most of the cases.  I agree though that 5-10s is too long.