@dgryski 2015-06-15
Looking at my tests over 20 runs, the windowsize=1 case seems to vary by +/- 12%.  That's significant variance, and common when microbenchmarking very fast routines.  System noise can throw off the results.