@dgryski 2015-06-15
If you're happy with this, I can push it to master or  you can hit the green merge button.